+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # gradle-pitest-plugin changelog
 
+## 1.9.0 - Unreleased
+
+ - PIT 1.9.4 by default (requires pitest-junit5-plugin 1.0.0!)
+ - Deprecate support for JDK 8 - [#299](https://github.com/szpak/gradle-pitest-plugin/issues/299)
+ - Support for `inputEncoding` and `outputEncoding` (named in plugin `inputCharset` and `outputCharset`) - [#310](https://github.com/szpak/gradle-pitest-plugin/issues/310)
+ - Add pitest dependencies only during configuration resolution - [#313](https://github.com/szpak/gradle-pitest-plugin/issues/313) - PR by [Ian O'Malley](https://github.com/omalleyian)
+ - Improve build cache hit cache with relative path - [#315](https://github.com/szpak/gradle-pitest-plugin/issues/315) - PR by [Siddardha Bezawada](https://github.com/SidB3)
+ - Remove (optional) `dependencyDistance` configuration parameter removed in PIT 1.9.0
+
+
+**Compatibility notes**
+1. Due to the plugin API changes in PIT 1.9.0, the new pitest-junit5-plugin 1.0.0 is required (other PIT plugins - if used - also need to be upgraded).
+2. The plugin is released with JDK 11. Source/class compatibility is still set to JDK 8, but this might change without notice.
+
+
 ## 1.7.4 - 2022-03-12
 
  - Do not pass `--testPlugin` for PIT 1.6.7+ for forward compatibility - [#277](https://github.com/szpak/gradle-pitest-plugin/issues/277)